backend/requirements.txt
fastapi
uvicorn
websockets
azure-communication-callautomation
azure-communication-identity
azure-cognitiveservices-speech
pydantic
python-multipart
pydub
numpy
aiohttp
python-dotenv
audioop-lts
openai
pandas
azure-search-documents
openpyxl
tabulate